{{cookiecutter.project_slug}}/requirements/base.txt
fastapi==0.63.0
python-multipart==0.0.5
pydantic==1.8.1
sqlalchemy==1.4.8
alembic==1.5.8
psycopg2-binary==2.8.6
python-jose==3.2.0
passlib==1.7.4
bcrypt==3.2.0

# Testing
# ------------------------------------------------------------------------------
pytest==6.2.3
Faker==8.1.0
tox==3.23.0

# Code quality
# ------------------------------------------------------------------------------
flake8==3.9.1
isort==5.8.0
coverage==5.5
black==20.8b1
mypy==0.812
sqlalchemy-stubs==0.4
